refactor: split task result/inspect/error from apistructs
