+++ apistructs/pipeline_task.go
@@ -15,12 +15,9 @@
 package apistructs
 
 import (
-	"fmt"
-	"sort"
-	"strings"
 	"time"
 
-	"github.com/erda-project/erda/pkg/metadata"
+	"github.com/erda-project/erda/internal/tools/pipeline/pkg/taskresult"
 )
 
 const (
@@ -32,7 +29,6 @@ const (
 	MSPTerminusOrgNameTag        = "msp.erda.cloud/org_name"
 	PipelineTaskMaxRetryLimit    = 144
 	PipelineTaskMaxRetryDuration = 24 * time.Hour
-	PipelineTaskMaxErrorPerHour  = 180
 )
 
 type PipelineTaskDTO struct {
@@ -40,13 +36,13 @@ type PipelineTaskDTO struct {
 	PipelineID uint64 `json:"pipelineID"`
 	StageID    uint64 `json:"stageID"`
 
-	Name   string             `json:"name"`
-	OpType string             `json:"opType"`         // get, put, task
-	Type   string             `json:"type,omitempty"` // git, buildpack, release, dice ... 当 OpType 为自定义任务时为空
-	Status PipelineStatus     `json:"status"`
-	Extra  PipelineTaskExtra  `json:"extra"`
-	Labels map[string]string  `json:"labels"`
-	Result PipelineTaskResult `json:"result"`
+	Name   string                        `json:"name"`
+	OpType string                        `json:"opType"`         // get, put, task
+	Type   string                        `json:"type,omitempty"` // git, buildpack, release, dice ... 当 OpType 为自定义任务时为空
+	Status PipelineStatus                `json:"status"`
+	Extra  PipelineTaskExtra             `json:"extra"`
+	Labels map[string]string             `json:"labels"`
+	Result taskresult.PipelineTaskResult `json:"result"`
 
 	IsSnippet             bool                       `json:"isSnippet"`
 	SnippetPipelineID     *uint64                    `json:"snippetPipelineID,omitempty"`
@@ -69,22 +65,6 @@ type PipelineTaskExtra struct {
 type TaskContainer struct {
 	TaskName    string `json:"taskName"`
 	ContainerID string `json:"containerID"`
-}
-
-// PipelineTaskResult spec.pipeline task only use metadata, task dto has all fields
-type PipelineTaskResult struct {
-	Metadata    metadata.Metadata          `json:"metadata,omitempty"`
-	Errors      []*PipelineTaskErrResponse `json:"errors,omitempty"`
-	MachineStat *PipelineTaskMachineStat   `json:"machineStat,omitempty"`
-	Inspect     string                     `json:"inspect,omitempty"`
-	Events      string                     `json:"events,omitempty"`
-}
-
-type PipelineTaskInspect struct {
-	Errors      []*PipelineTaskErrResponse `json:"errors,omitempty"`
-	MachineStat *PipelineTaskMachineStat   `json:"machineStat,omitempty"`
-	Inspect     string                     `json:"inspect,omitempty"`
-	Events      string                     `json:"events,omitempty"`
 }
 
 type PipelineTaskSnippetDetail struct {
@@ -110,48 +90,6 @@ type PipelineTaskGetBootstrapInfoResponse struct {
 
 type PipelineTaskGetBootstrapInfoResponseData struct {
 	Data []byte `json:"data"`
-}
-
-type PipelineTaskMachineStat struct {
-	Host PipelineTaskMachineHostStat `json:"host,omitempty"`
-	Pod  PipelineTaskMachinePodStat  `json:"pod,omitempty"`
-	Load PipelineTaskMachineLoadStat `json:"load,omitempty"`
-	Mem  PipelineTaskMachineMemStat  `json:"mem,omitempty"`
-	Swap PipelineTaskMachineSwapStat `json:"swap,omitempty"`
-}
-type PipelineTaskMachineHostStat struct {
-	HostIP          string `json:"hostIP,omitempty"`
-	Hostname        string `json:"hostname,omitempty"`
-	UptimeSec       uint64 `json:"uptimeSec,omitempty"`
-	BootTimeSec     uint64 `json:"bootTimeSec,omitempty"`
-	OS              string `json:"os,omitempty"`
-	Platform        string `json:"platform,omitempty"`
-	PlatformVersion string `json:"platformVersion,omitempty"`
-	KernelVersion   string `json:"kernelVersion,omitempty"`
-	KernelArch      string `json:"kernelArch,omitempty"`
-}
-type PipelineTaskMachinePodStat struct {
-	PodIP string `json:"podIP,omitempty"`
-}
-type PipelineTaskMachineLoadStat struct {
-	Load1  float64 `json:"load1,omitempty"`
-	Load5  float64 `json:"load5,omitempty"`
-	Load15 float64 `json:"load15,omitempty"`
-}
-type PipelineTaskMachineMemStat struct { // all byte
-	Total       uint64  `json:"total,omitempty"`
-	Available   uint64  `json:"available,omitempty"`
-	Used        uint64  `json:"used,omitempty"`
-	Free        uint64  `json:"free,omitempty"`
-	UsedPercent float64 `json:"usedPercent,omitempty"`
-	Buffers     uint64  `json:"buffers,omitempty"`
-	Cached      uint64  `json:"cached,omitempty"`
-}
-type PipelineTaskMachineSwapStat struct { // all byte
-	Total       uint64  `json:"total,omitempty"`
-	Used        uint64  `json:"used,omitempty"`
-	Free        uint64  `json:"free,omitempty"`
-	UsedPercent float64 `json:"usedPercent,omitempty"`
 }
 
 const TaskLoopTimeBegin = 1
@@ -192,104 +130,6 @@ func (l *PipelineTaskLoop) IsEmpty() bool {
 	}
 
 	return true
-}
-
-type PipelineTaskErrResponse struct {
-	Code string             `json:"code"`
-	Msg  string             `json:"msg"`
-	Ctx  PipelineTaskErrCtx `json:"ctx"`
-}
-
-type PipelineTaskErrCtx struct {
-	StartTime time.Time `json:"startTime"`
-	EndTime   time.Time `json:"endTime"`
-	Count     uint64    `json:"count"`
-}
-
-func (c *PipelineTaskErrCtx) CalculateFrequencyPerHour() uint64 {
-	if c.StartTime.IsZero() || c.EndTime.IsZero() || c.EndTime.Sub(c.StartTime) <= time.Hour {
-		return c.Count
-	}
-	return uint64(float64(c.Count) / c.EndTime.Sub(c.StartTime).Hours())
-}
-
-type orderedResponses []*PipelineTaskErrResponse
-
-func (o orderedResponses) Len() int           { return len(o) }
-func (o orderedResponses) Less(i, j int) bool { return o[i].Ctx.EndTime.Before(o[j].Ctx.EndTime) }
-func (o orderedResponses) Swap(i, j int)      { o[i], o[j] = o[j], o[i] }
-
-func (t *PipelineTaskInspect) AppendError(newResponses ...*PipelineTaskErrResponse) []*PipelineTaskErrResponse {
-	if len(newResponses) == 0 {
-		return t.Errors
-	}
-	var orderd orderedResponses
-	for _, g := range t.Errors {
-		orderd = append(orderd, g)
-	}
-
-	var newResponseOrder orderedResponses
-	now := time.Now()
-	for index, g := range newResponses {
-		if g.Ctx.StartTime.IsZero() {
-			g.Ctx.StartTime = now.Add(time.Duration(index) * time.Millisecond)
-		}
-		if g.Ctx.EndTime.IsZero() {
-			g.Ctx.EndTime = now.Add(time.Duration(index) * time.Millisecond)
-		}
-		if g.Ctx.Count == 0 {
-			g.Ctx.Count = 1
-		}
-		newResponseOrder = append(newResponseOrder, g)
-	}
-	sort.Sort(newResponseOrder)
-
-	var lastResponse *PipelineTaskErrResponse
-	if len(orderd) != 0 {
-		lastResponse = orderd[len(orderd)-1]
-	}
-
-	for _, g := range newResponseOrder {
-		if lastResponse == nil {
-			orderd = append(orderd, g)
-			lastResponse = g
-			continue
-		}
-
-		if strings.EqualFold(lastResponse.Msg, g.Msg) {
-			if !g.Ctx.StartTime.IsZero() && g.Ctx.StartTime.Before(lastResponse.Ctx.StartTime) {
-				lastResponse.Ctx.StartTime = g.Ctx.StartTime
-			}
-			if g.Ctx.EndTime.After(lastResponse.Ctx.EndTime) {
-				lastResponse.Ctx.EndTime = g.Ctx.EndTime
-			}
-			lastResponse.Ctx.Count++
-			continue
-		} else {
-			orderd = append(orderd, g)
-			lastResponse = g
-		}
-	}
-	return orderd
-}
-
-func (t *PipelineTaskInspect) ConvertErrors() {
-	for _, response := range t.Errors {
-		if response.Ctx.Count > 1 {
-			response.Msg = fmt.Sprintf("%s\nstartTime: %s\nendTime: %s\ncount: %d",
-				response.Msg, response.Ctx.StartTime.Format("2006-01-02 15:04:05"),
-				response.Ctx.EndTime.Format("2006-01-02 15:04:05"), response.Ctx.Count)
-		}
-	}
-}
-
-func (t *PipelineTaskInspect) IsErrorsExceed() (bool, *PipelineTaskErrResponse) {
-	for _, g := range t.Errors {
-		if g.Ctx.CalculateFrequencyPerHour() > PipelineTaskMaxErrorPerHour {
-			return true, g
-		}
-	}
-	return false, nil
 }
 
 func (l *PipelineTaskLoop) Duplicate() *PipelineTaskLoop {
